Refactor theme management: replace old theme provider with new implementation and restore ModeToggle component
